+++ database/src/main/java/com/cookguide/database/cookAPI/application/dto/request/RecipesRequestDTO.java
@@ -1,0 +1,4 @@
+package com.cookguide.database.cookAPI.application.dto.request;
+
+public class RecipesRequestDTO {
+}
