+++ database/src/main/java/com/cookguide/database/cookAPI/application/dto/response/AccountResponseDTO.java
@@ -1,0 +1,16 @@
+package com.cookguide.database.cookAPI.application.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class AccountResponseDTO {
+    private Long uid;
+    private String username;
+    
+}
