+++ database/src/main/java/com/cookguide/database/cookAPI/application/dto/response/RecipesResponseDTO.java
@@ -1,0 +1,4 @@
+package com.cookguide.database.cookAPI.application.dto.response;
+
+public class RecipesResponseDTO {
+}
