+++ database/src/main/java/com/cookguide/database/cookAPI/application/services/HealthInfoService.java
@@ -1,0 +1,4 @@
+package com.cookguide.database.cookAPI.application.services;
+
+public interface HealthInfoService {
+}
