+++ database/src/main/java/com/cookguide/database/cookAPI/application/services/PreferenceService.java
@@ -1,0 +1,8 @@
+package com.cookguide.database.cookAPI.application.services;
+
+import com.cookguide.database.cookAPI.domain.entities.Preference;
+
+public interface PreferenceService {
+    public abstract Preference createPreference(Preference preference);
+
+}
