+++ database/src/main/java/com/cookguide/database/cookAPI/application/services/impl/HealthInfoServiceImpl.java
@@ -1,0 +1,4 @@
+package com.cookguide.database.cookAPI.application.services.impl;
+
+public class HealthInfoServiceImpl {
+}
