+++ database/src/main/java/com/cookguide/database/cookAPI/infraestructure/repositories/PreferenceRepository.java
@@ -1,0 +1,8 @@
+package com.cookguide.database.cookAPI.infraestructure.repositories;
+
+import com.cookguide.database.cookAPI.domain.entities.Preference;
+import org.springframework.data.repository.CrudRepository;
+
+public interface PreferenceRepository extends CrudRepository<Preference, Integer>
+{
+}
