+++ database/src/main/java/com/cookguide/database/cookAPI/infraestructure/repositories/RecipesRepository.java
@@ -1,0 +1,7 @@
+package com.cookguide.database.cookAPI.infraestructure.repositories;
+
+import com.cookguide.database.cookAPI.domain.entities.Recipes;
+import org.springframework.data.repository.CrudRepository;
+
+public interface RecipesRepository extends CrudRepository<Recipes,Integer > {
+}
