--- database/src/main/java/com/cookguide/database/exception/AccountNotFoundException.java
@@ -1,5 +1,0 @@
-package com.cookguide.database.exception;
-
-public class AccountNotFoundException extends RuntimeException{
-    private static final long serialVersionUID = 1L;
-}
