--- database/src/main/java/com/cookguide/database/exception/ErrorMessage.java
@@ -1,19 +1,0 @@
-package com.cookguide.database.exception;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class ErrorMessage {
-    private int statusCode;
-    private String message;
-    private String description;
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-yyyy hh:mm:ss")
-    private LocalDateTime timeStamp;
-}
