--- database/src/main/java/com/cookguide/database/exception/ResourceNotFoundException.java
@@ -1,11 +1,0 @@
-package com.cookguide.database.exception;
-
-public class ResourceNotFoundException extends RuntimeException {
-    public ResourceNotFoundException(){
-        super();
-    }
-
-    public ResourceNotFoundException(String message){
-        super(message);
-    }
-}
