--- database/src/main/java/com/cookguide/database/exception/ValidationException.java
@@ -1,11 +1,0 @@
-package com.cookguide.database.exception;
-
-public class ValidationException extends RuntimeException {
-    public ValidationException(){
-        super();
-    }
-
-    public ValidationException(String message){
-        super(message);
-    }
-}
