--- database/src/main/java/com/cookguide/database/repository/HealthInfoRepository.java
@@ -1,8 +1,0 @@
-package com.cookguide.database.repository;
-
-import com.cookguide.database.model.Health_info;
-import org.springframework.data.repository.CrudRepository;
-
-public interface HealthInfoRepository extends CrudRepository<Health_info, Integer> {
-
-}
