--- database/src/main/java/com/cookguide/database/repository/RecipesRepository.java
@@ -1,7 +1,0 @@
-package com.cookguide.database.repository;
-
-import com.cookguide.database.model.Recipes;
-import org.springframework.data.repository.CrudRepository;
-
-public interface RecipesRepository extends CrudRepository<Recipes,Integer > {
-}
