--- database/src/main/java/com/cookguide/database/service/HealthInfoService.java
@@ -1,4 +1,0 @@
-package com.cookguide.database.service;
-
-public interface HealthInfoService {
-}
