--- database/src/main/java/com/cookguide/database/service/PreferenceService.java
@@ -1,8 +1,0 @@
-package com.cookguide.database.service;
-
-import com.cookguide.database.model.Preference;
-
-public interface PreferenceService {
-    public abstract Preference createPreference(Preference preference);
-
-}
