--- database/src/main/java/com/cookguide/database/service/RecipesService.java
@@ -1,17 +1,0 @@
-package com.cookguide.database.service;
-
-import com.cookguide.database.model.Recipes;
-
-import java.util.List;
-
-public interface RecipesService {
-    public Recipes createRecipes(Recipes recipes);
-
-    Recipes updateRecipes(Recipes recipes);
-
-    List<Recipes> getAllRecipes();
-
-    void deleteRecipes(int id);
-
-    boolean isRecipesExist(int id);
-}
