--- database/src/main/java/com/cookguide/database/service/impl/HealthInfoServiceImpl.java
@@ -1,4 +1,0 @@
-package com.cookguide.database.service.impl;
-
-public class HealthInfoServiceImpl {
-}
