+++ database/src/main/java/com/cookguide/database/shared/config/ModelMapperConfig.java
@@ -1,0 +1,13 @@
+package com.cookguide.database.shared.config;
+
+import org.modelmapper.ModelMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class ModelMapperConfig {
+    @Bean
+    public ModelMapper modelMapper() {
+        return new ModelMapper();
+    }
+}
