+++ database/src/main/java/com/cookguide/database/shared/exception/AccountNotFoundException.java
@@ -1,0 +1,5 @@
+package com.cookguide.database.shared.exception;
+
+public class AccountNotFoundException extends RuntimeException{
+    private static final long serialVersionUID = 1L;
+}
