+++ database/src/main/java/com/cookguide/database/shared/exception/ResourceNotFoundException.java
@@ -1,0 +1,11 @@
+package com.cookguide.database.shared.exception;
+
+public class ResourceNotFoundException extends RuntimeException {
+    public ResourceNotFoundException(){
+        super();
+    }
+
+    public ResourceNotFoundException(String message){
+        super(message);
+    }
+}
