+++ database/src/main/java/com/cookguide/database/shared/exception/ValidationException.java
@@ -1,0 +1,11 @@
+package com.cookguide.database.shared.exception;
+
+public class ValidationException extends RuntimeException {
+    public ValidationException(){
+        super();
+    }
+
+    public ValidationException(String message){
+        super(message);
+    }
+}
