+++ database/src/main/java/com/cookguide/database/shared/model/enums/Estatus.java
@@ -1,0 +1,6 @@
+package com.cookguide.database.shared.model.enums;
+
+public enum Estatus {
+    SUCCESS,
+    ERROR
+}
